11.26 string函数

+++ chapter-8/lib/string.h
@@ -1,5 +1,6 @@
 #ifndef __LIB_STRING_H
 #define __LIB_STRING_H  
+#include "stdint.h"                 //我们的stdint.h中定义了数据类型，包含进来
 
 void* memset(void* dst_, uint8_t value, uint32_t size);
 void memcpy(void* dst_, const void* src_, uint32_t size);
@@ -10,5 +11,5 @@ char* strcat(char* dst_, const char* src_);
 char* strchr(const char* str, const uint8_t ch);
 char* strrchr(const char* str, const uint8_t ch);
 uint32_t strchrs(const char* str, const uint8_t ch);
-int strcmp(const char* a_, const char* b_);
+int8_t strcmp(const char* a_, const char* b_);
 #endif
